ai_engine/requirements.txt
# ─────────────────────────────────────────────────────────
#  Lucid AI — Agent Engine Dependencies
# ─────────────────────────────────────────────────────────

# Core Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
websockets>=12.0
python-multipart>=0.0.6

# Authentication
python-jose[cryptography]>=3.3.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# LLM Providers
google-generativeai>=0.8.0
litellm>=1.50.0

# Docker (for Docker host management)
docker>=7.1.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0

# ─────────────────────────────────────────────────────────
#  OpenHands SDK V1 (optional — not yet published on PyPI)
#  Install fails gracefully; app runs in mock mode without them.
#  Uncomment when packages are available:
# ─────────────────────────────────────────────────────────
# openhands-sdk>=1.0.0
# openhands-tools>=1.0.0
# openhands-workspace>=1.0.0
# openhands-agent-server>=1.0.0
